requirements.txt
altgraph==0.17.4
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
cssselect==1.2.0
cssutils==2.11.1
google-api-core==2.20.0
google-api-python-client==2.146.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
gspread==6.1.2
httplib2==0.22.0
idna==3.10
lxml==5.3.0
more-itertools==10.5.0
numpy==2.1.1
oauth2client==3.0.0
oauthlib==3.2.2
packaging==24.1
pandas==2.2.3
premailer==3.10.0
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
pymssql==2.3.1
pyodbc==5.1.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.3
yagmail==0.15.293
